refactor: Celery config file moved to settings root folder
