+++ Project/__init__.py
@@ -1,6 +1,6 @@
 import pymysql
 
-from .settings.celery_worker.worker import app as celery_app
+from .settings.celery_worker import app as celery_app
 
 
 pymysql.install_as_MySQLdb()
