+++ Project/settings/celery_worker.py
