+++ Project/settings/jet_settings.py
@@ -1,3 +1,5 @@
+import os
+
 from django.utils.html import format_html
 
 from Project.settings.labels_with_svg import black_list_with_icon
@@ -11,6 +13,17 @@ from Project.settings.labels_with_svg import suggestion_label_with_icon
 from Project.settings.labels_with_svg import swagger_label_with_icon
 from Project.settings.labels_with_svg import user_label_with_icon
 
+
+if not os.environ.get("PROJECT_URL"):
+    os.environ.setdefault("PROJECT_URL", "localhost")
+if not os.environ.get("GRAFANA_URL"):
+    os.environ.setdefault("GRAFANA_URL", f"{os.environ['PROJECT_URL']}:3000/")
+if not os.environ.get("PROMETHEUS_URL"):
+    os.environ.setdefault(
+        "PROMETHEUS_URL", f"{os.environ['PROJECT_URL']}:9090/"
+    )
+if not os.environ.get("FLOWER_URL"):
+    os.environ.setdefault("FLOWER_URL", f"{os.environ['PROJECT_URL']}:5555/")
 
 """
 JET Documentation: https://django-jet-reboot.readthedocs.io/
@@ -63,7 +76,7 @@ JET_SIDE_MENU_ITEMS: list = [
         ],
     },
     {
-        "label": ("Adminitsration"),
+        "label": ("Administration"),
         "app_label": "admin",
         "items": [
             {"name": "logentry", "label": format_html(log_label_with_icon)}
@@ -89,27 +102,17 @@ JET_SIDE_MENU_ITEMS: list = [
         "items": [
             {
                 "label": "Grafana",
-                "url": "http://localhost:3000/",
+                "url": f"http://{os.environ['GRAFANA_URL']}",
                 "url_blank": True,
             },
             {
                 "label": "Prometheus",
-                "url": "http://localhost:9090/",
-                "url_blank": True,
-            },
-            {
-                "label": "RabbitMQ",
-                "url": "http://localhost:15672/",
+                "url": f"http://{os.environ['PROMETHEUS_URL']}",
                 "url_blank": True,
             },
             {
                 "label": "Flower",
-                "url": "http://localhost:5555/",
-                "url_blank": True,
-            },
-            {
-                "label": "CloudBeaver",
-                "url": "http://localhost:8081/",
+                "url": f"http://{os.environ['FLOWER_URL']}",
                 "url_blank": True,
             },
         ],
